feature: Applied the custom aliases for numeric types

+++ core/graphics/animations/animation.cpp
@@ -3,6 +3,8 @@
 #include <algorithm>
 #include <utility>
 
+#include "aporia_types.hpp"
+
 namespace Aporia
 {
     Animation::Animation(std::string name, std::vector<AnimationFrame>&& frames)
@@ -17,14 +19,14 @@ namespace Aporia
     void Animation::play()
     {
         _current_frame = 0;
-        _increment = [](size_t curr_frame, size_t frame_count){ return (curr_frame + 1) % frame_count; };
+        _increment = [](u64 curr_frame, u64 frame_count){ return (curr_frame + 1) % frame_count; };
         _timer.reset();
     }
 
     void Animation::play_once()
     {
         _current_frame = 0;
-        _increment = [](size_t curr_frame, size_t frame_count){ return std::min(curr_frame + 1, frame_count - 1); };
+        _increment = [](u64 curr_frame, u64 frame_count){ return std::min(curr_frame + 1, frame_count - 1); };
         _timer.reset();
     }
