+++ core/scene_manager.hpp
@@ -1,8 +1,9 @@
 #pragma once
 
-#include <cstddef>
 #include <memory>
 #include <vector>
+
+#include "aporia_types.hpp"
 
 namespace Aporia
 {
@@ -16,20 +17,20 @@ namespace Aporia
     public:
         SceneManager();
 
-        size_t add(Ref scene);
+        u64 add(Ref scene);
         void remove(Ref scene);
 
-        void load_scene(size_t id);
+        void load_scene(u64 id);
         void load_scene(Ref scene);
 
         Ref get_current_scene() const { return _scenes[_current_scene]; };
 
     private:
-        size_t _index(Ref scene);
-        bool _load(size_t id);
+        u64 _index(Ref scene);
+        bool _load(u64 id);
 
     private:
         Container _scenes;
-        size_t _current_scene = 0;
+        u64 _current_scene = 0;
     };
 }
